// src/app/interfaces/common/ten-points.interface.ts
export interface TenPoints {
  _id?: string;
  name?: string;
  nameEn?: string;
  image?: string;
  seoImage?: string;
  images?: string[];
  informations?: any[];
  description?: string;
  isHtml?: boolean;
  descriptionEn?: string;
  isHtmlEn?: boolean;
  websiteUrl?: string;
  slug?: string;
  createdAt?: Date;
  updatedAt?: Date;
}